Clean up code

+++ source/effect_expression.hpp
@@ -265,7 +265,7 @@ namespace reshadefx
 	{
 		std::string name;
 		type type;
-		uint32_t member_index = 0;
+		size_t member_index = 0;
 		uint32_t offset = 0;
 		std::unordered_map<std::string, std::pair<struct type, constant>> annotations;
 		bool has_initializer_value = false;
